client/modules/About & client/modules/Legal: clean up Proptypes
